% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parse_styles}
\alias{parse_styles}
\title{Parse style elements}
\usage{
parse_styles(familyframe)
}
\arguments{
\item{familyframe}{A frame derived from \code{\link[systemfonts:system_fonts]{systemfonts::system_fonts()}}}
}
\value{
A list of form: list(regular=c(path1, ...), bold=c(path2, ...), italic=c(path3, ...), bolditali=c(path3, ...))
}
\description{
Parse style elements
}
